extract chat message menu to a component so it's easier to refactor
